Moved util functions to closest scopes
